+++ src/util/main/sp/it/util/file/FileExtensions.kt
@@ -1,11 +1,13 @@
 package sp.it.util.file
 
 import java.io.File
+import java.io.IOException
 import java.nio.file.FileVisitOption
 import java.nio.file.Files
 import java.nio.file.Path
 import java.nio.file.attribute.BasicFileAttributes
 import kotlin.io.path.pathString
+import sp.it.util.dev.Blocks
 import sp.it.util.file.FileType.*
 import sp.it.util.functional.orNull
 import sp.it.util.functional.runTry
@@ -29,3 +31,12 @@ fun File.getFilesR(maxDepth: Int = Int.MAX_VALUE, fileType: FileType?, vararg op
    }
    return runTry { Files.find(toPath(), maxDepth, p, *options).use { it.map(m).toList() } }.orNull().orEmpty()
 }
+
+@Blocks @Throws(IOException::class)
+fun File.deleteOrThrow(): Boolean = delete().also { if (!it) throw IOException("Failed to delete file=$this") }
+
+@Blocks @Throws(IOException::class)
+fun File.deleteRecursivelyOrThrow(): Boolean = delete().also { if (!it) throw IOException("Failed to delete file=$this") }
+
+@Blocks @Throws(IOException::class)
+fun File.setExecutableOrThrow(executable: Boolean): Boolean = setExecutable(executable).also { if (!it) throw IOException("Failed to set executable=$executable for file=$this") }